workers/requirements.txt
httpx==0.27.0
supabase==2.3.0
python-dotenv==1.0.0
playwright==1.42.0
tenacity==8.2.3
